requirements.txt
pytest
pyqt5
tensorflow==1.12
opencv-python
numpy
pillow